3D planning in progress
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@4879 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/sbpl/src/discrete_space_information/nav3dkin/environment_nav3Dkin.cpp
@@ -1039,7 +1039,7 @@ void EnvironmentNAV3DKIN::GetPredsofChangedEdges(vector<nav2dcell_t>* changedcel
 {
 	nav2dcell_t cell;
 
-	for(int i = 0; i < changedcellsV->size(); i++)
+	for(int i = 0; i < (int)changedcellsV->size(); i++)
 	{
 		cell = changedcellsV->at(i);
 		for(int tind = 0; tind < NAV3DKIN_THETADIRS; tind++)
